Add Maintenance Guard Component separate from AuthGuard to fix login rendering
