+++ frontend/src/authentication/components/AuthGuard.js
@@ -1,59 +1,18 @@
 import PropTypes from 'prop-types';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { Navigate, useLocation } from 'react-router-dom';
 import { Login } from '../views/Login';
 import { useAuth } from '../hooks';
 import {
-  isModuleEnabled,
-  ModuleNames,
   RegexToValidateWindowPathName,
   WindowPathLengthThreshold
-} from 'utils';
-import { useClient, useGroups } from 'services';
-import { LoadingScreen, NoAccessMaintenanceWindow } from 'design';
-import { getMaintenanceStatus } from '../../modules/Maintenance/services';
-import {
-  PENDING_STATUS,
-  ACTIVE_STATUS
-} from '../../modules/Maintenance/components/MaintenanceViewer';
-import { SET_ERROR, useDispatch } from 'globalErrors';
+} from '../../utils';
 
 export const AuthGuard = (props) => {
   const { children } = props;
   const auth = useAuth();
   const location = useLocation();
   const [requestedLocation, setRequestedLocation] = useState(null);
-  const [isNoAccessMaintenance, setNoAccessMaintenanceFlag] = useState(null);
-  const client = useClient();
-  const groups = useGroups();
-  const dispatch = useDispatch();
-
-  const checkMaintenanceMode = async () => {
-    const response = await client.query(getMaintenanceStatus());
-    if (!response.errors && response.data.getMaintenanceWindowStatus != null) {
-      if (
-        [PENDING_STATUS, ACTIVE_STATUS].includes(
-          response.data.getMaintenanceWindowStatus.status
-        ) &&
-        response.data.getMaintenanceWindowStatus.mode === 'NO-ACCESS' &&
-        !groups.includes('DAAdministrators')
-      ) {
-        setNoAccessMaintenanceFlag(true);
-      } else {
-        setNoAccessMaintenanceFlag(false);
-      }
-    }
-  };
-
-  useEffect(async () => {
-    // Check if the maintenance window is enabled and has NO-ACCESS Status
-    // If yes then display a blank screen with a message that data.all is in maintenance mode ( Check use of isNoAccessMaintenance state )
-    if (isModuleEnabled(ModuleNames.MAINTENANCE) === true) {
-      if (client) {
-        checkMaintenanceMode().catch((e) => dispatch({ type: SET_ERROR, e }));
-      }
-    }
-  }, [client, groups]);
 
   if (!auth.isAuthenticated) {
     if (location.pathname !== requestedLocation) {
@@ -70,17 +29,6 @@ export const AuthGuard = (props) => {
     }
 
     return <Login />;
-  }
-
-  if (
-    isNoAccessMaintenance == null &&
-    isModuleEnabled(ModuleNames.MAINTENANCE) === true
-  ) {
-    return <LoadingScreen />;
-  }
-
-  if (isNoAccessMaintenance === true) {
-    return <NoAccessMaintenanceWindow />;
   }
 
   if (requestedLocation && location.pathname !== requestedLocation) {
