+++ frontend/src/authentication/components/index.js
@@ -1,3 +1,4 @@
 export * from './AuthGuard';
 export * from './GuestGuard';
+export * from './MaintenanceGuard';
 export * from './LoginButton';
